+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
 ## Unreleased
+- make madlib a class
 
 ## [[0.1]]()